refactor: untangle repo imports, further simplify store service
