+++ src/database/entity/Order.ts
@@ -1,5 +1,4 @@
 import {
-    BeforeInsert,
     Column,
     Entity,
     JoinTable,
@@ -7,10 +6,9 @@ import {
     ManyToOne,
     PrimaryGeneratedColumn,
 } from 'typeorm'
-import { productRepo } from '../../index.js'
+
 import { Customer } from './Customer.js'
 import { Product } from './Product.js'
-import { calculateOrderTotal } from '../../utils/calculateOrderTotal.js'
 
 type Status = 'pending' | 'complete'
 
@@ -32,31 +30,7 @@ export class Order {
     @ManyToOne(() => Customer, (customer) => customer.orders)
     owner: Customer
 
-    // @BeforeInsert()
-    // async setTotal() {
-    //     let total = 0
-    //     for (const itemId of this.cartItemIds) {
-    //         let productInfo = await productRepo.findOneBy({ id: itemId })
-    //         if (!productInfo) {
-    //             throw new Error(
-    //                 `Couldn't find any info for product with ID ${itemId} - order total was not calculated`,
-    //             )
-    //         }
-    //         total += productInfo.price
-    //     }
-    //     this.total = total
-    // }
-
-
-    *[Symbol.iterator]() {
-        for (const item of this.products) {
-            yield item
-        }
-    }
-
-
     constructor(owner: Customer) {
         this.owner = owner
-
     }
 }
