+++ src/utils/reduceProductQuantity.ts
@@ -1,9 +1,7 @@
-
-import type { Order } from "../database/entity/Order.js";
-import { productRepo } from "../index.js";
+import type { Order } from '../database/entity/Order.js'
+import { productRepo } from '../database/connection.js'
 
 export const reduceProductQuantity = async (orders: Order[]) => {
-
     for (const order of orders) {
         for (const product of order.products) {
             product.quantityInStock -= 1
@@ -11,6 +9,4 @@ export const reduceProductQuantity = async (orders: Order[]) => {
             console.log(`Product ${product.id}'s quantity was reduced`)
         }
     }
-
-
 }
